refactor(enrollment): remove legacy `AcademicInfoStep` and streamline academic info widget structure
