+++ lib/features/enrollment/presentation/widgets/academic_info/academic_info_widgets.dart
@@ -1,4 +1,5 @@
-export 'academic_info_step_body.dart';
+export 'previous_academic_info_step_body.dart';
+export 'target_academic_info_step_body.dart';
 export 'academic_info_card.dart';
 export 'dropdown_field.dart';
 export 'previous_year_fields.dart';
