+++ lib/features/enrollment/presentation/widgets/previous_academic_info_step.dart
@@ -3,8 +3,6 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:school_app_flutter/core/di/injection.dart';
 import 'package:school_app_flutter/core/widgets/app_snack_bar.dart';
 import 'package:school_app_flutter/features/academic_year/presentation/bloc/enrollment_academic_info_bloc.dart';
-import 'package:school_app_flutter/features/bootstrap/domain/entities/bootstrap.dart';
-import 'package:school_app_flutter/features/bootstrap/presentation/bloc/bootstrap_bloc.dart';
 import 'package:school_app_flutter/features/enrollment/domain/entities/enrollment_school_detail.dart';
 import 'package:school_app_flutter/features/enrollment/presentation/bloc/enrollment_bloc.dart';
 import 'package:school_app_flutter/features/enrollment/presentation/bloc/enrollment_stepper_flow_bloc.dart';
@@ -28,7 +26,8 @@ class PreviousAcademicInfoStep extends StatefulWidget {
   });
 
   @override
-  State<PreviousAcademicInfoStep> createState() => PreviousAcademicInfoStepState();
+  State<PreviousAcademicInfoStep> createState() =>
+      PreviousAcademicInfoStepState();
 }
 
 class PreviousAcademicInfoStepState extends State<PreviousAcademicInfoStep> {
@@ -38,14 +37,10 @@ class PreviousAcademicInfoStepState extends State<PreviousAcademicInfoStep> {
   late final TextEditingController _prevLevelController;
   late final TextEditingController _prevRateController;
   late final TextEditingController _prevRankController;
-  late final TextEditingController _currYearController;
-  late final TextEditingController _targetOptionController;
 
   late bool _validatedPreviousYear;
-  late String _selectedSchoolLevelGroupId;
-  late String _selectedSchoolLevelId;
+
   String _academicYearId = '';
-  bool _bootstrapDefaultsApplied = false;
 
   String _initialPrevYear = '';
   String _initialPrevSchool = '';
@@ -53,11 +48,7 @@ class PreviousAcademicInfoStepState extends State<PreviousAcademicInfoStep> {
   String _initialPrevLevel = '';
   String _initialPrevRate = '';
   String _initialPrevRank = '';
-  String _initialCurrYear = '';
-  String _initialAcademicYearId = '';
   bool _initialValidatedPreviousYear = false;
-  String _initialSchoolLevelGroupId = '';
-  String _initialSchoolLevelId = '';
 
   bool _isDirty = false;
   bool _isValid = false;
@@ -66,6 +57,7 @@ class PreviousAcademicInfoStepState extends State<PreviousAcademicInfoStep> {
   bool _isHydratingFromDetail = false;
 
   bool get _canSave => _stepState.canSave;
+
   StepFormState get _stepState =>
       StepFormState(dirty: _isDirty, valid: _isValid, saving: _isSaving);
 
@@ -130,8 +122,6 @@ class PreviousAcademicInfoStepState extends State<PreviousAcademicInfoStep> {
     _prevLevelController = TextEditingController();
     _prevRateController = TextEditingController();
     _prevRankController = TextEditingController();
-    _currYearController = TextEditingController();
-    _targetOptionController = TextEditingController();
 
     _syncFromEnrollmentDetail(widget.enrollmentDetail, resetSnapshot: true);
 
@@ -163,8 +153,6 @@ class PreviousAcademicInfoStepState extends State<PreviousAcademicInfoStep> {
       _prevRankController.text = _normalizeRankFromInt(detail.previousRank);
       _academicYearId = detail.academicYearId;
       _validatedPreviousYear = detail.validatedPreviousYear;
-      _selectedSchoolLevelGroupId = detail.schoolLevelGroupId;
-      _selectedSchoolLevelId = detail.schoolLevelId;
     } finally {
       _isHydratingFromDetail = false;
     }
@@ -176,11 +164,7 @@ class PreviousAcademicInfoStepState extends State<PreviousAcademicInfoStep> {
       _initialPrevLevel = detail.previousSchoolLevel.trim();
       _initialPrevRate = _normalizeRateFromDouble(detail.previousRate);
       _initialPrevRank = _normalizeRankFromInt(detail.previousRank);
-      _initialCurrYear = _currYearController.text.trim();
-      _initialAcademicYearId = _academicYearId;
       _initialValidatedPreviousYear = detail.validatedPreviousYear;
-      _initialSchoolLevelGroupId = detail.schoolLevelGroupId;
-      _initialSchoolLevelId = detail.schoolLevelId;
     }
   }
 
@@ -191,11 +175,7 @@ class PreviousAcademicInfoStepState extends State<PreviousAcademicInfoStep> {
     _initialPrevLevel = _prevLevelController.text.trim();
     _initialPrevRate = _normalizeRate(_prevRateController.text);
     _initialPrevRank = _normalizeRank(_prevRankController.text);
-    _initialCurrYear = _currYearController.text.trim();
-    _initialAcademicYearId = _academicYearId;
     _initialValidatedPreviousYear = _validatedPreviousYear;
-    _initialSchoolLevelGroupId = _selectedSchoolLevelGroupId;
-    _initialSchoolLevelId = _selectedSchoolLevelId;
   }
 
   @override
@@ -203,7 +183,6 @@ class PreviousAcademicInfoStepState extends State<PreviousAcademicInfoStep> {
     super.didUpdateWidget(oldWidget);
     if (oldWidget.enrollmentDetail != widget.enrollmentDetail) {
       _syncFromEnrollmentDetail(widget.enrollmentDetail, resetSnapshot: true);
-      _bootstrapDefaultsApplied = false;
       _showValidationHints = false;
       _isSaving = false;
       _recomputeFormState(notifyParent: false);
@@ -228,8 +207,6 @@ class PreviousAcademicInfoStepState extends State<PreviousAcademicInfoStep> {
     _prevLevelController.dispose();
     _prevRateController.dispose();
     _prevRankController.dispose();
-    _currYearController.dispose();
-    _targetOptionController.dispose();
     _academicInfoBloc.close();
     super.dispose();
   }
@@ -342,63 +319,8 @@ class PreviousAcademicInfoStepState extends State<PreviousAcademicInfoStep> {
         previousRate: double.tryParse(_prevRateController.text.trim()) ?? 0.0,
         previousRank: int.tryParse(_prevRankController.text.trim()),
         validatedPreviousYear: _validatedPreviousYear,
-        schoolLevelId: _selectedSchoolLevelId,
-        schoolLevelGroupId: _selectedSchoolLevelGroupId,
       ),
     );
-  }
-
-  void _applyBootstrapDefaults(Bootstrap bootstrap) {
-    bool bootstrapChanged = false;
-    final previousAcademicYearId = _academicYearId;
-    final previousCurrentYear = _currYearController.text;
-    _academicYearId = bootstrap.currentAcademicYear.id;
-    _currYearController.text = bootstrap.currentAcademicYear.name;
-    if (previousAcademicYearId != _academicYearId ||
-        previousCurrentYear != _currYearController.text) {
-      bootstrapChanged = true;
-    }
-    if (_initialAcademicYearId.isEmpty) {
-      _initialAcademicYearId = _academicYearId;
-    }
-    if (_initialCurrYear.isEmpty) {
-      _initialCurrYear = _currYearController.text.trim();
-    }
-    if (_bootstrapDefaultsApplied) return;
-
-    final groupBundles = bootstrap.schoolLevelGroups;
-    final selectedGroupBundle = groupBundles
-        .where((g) => g.schoolLevelGroup.id == _selectedSchoolLevelGroupId)
-        .firstOrNull;
-
-    if (selectedGroupBundle == null && groupBundles.isNotEmpty) {
-      _selectedSchoolLevelGroupId = groupBundles.first.schoolLevelGroup.id;
-      _initialSchoolLevelGroupId = _selectedSchoolLevelGroupId;
-      bootstrapChanged = true;
-    }
-
-    final resolvedGroupBundle = groupBundles
-        .where((g) => g.schoolLevelGroup.id == _selectedSchoolLevelGroupId)
-        .firstOrNull;
-    if (resolvedGroupBundle != null &&
-        resolvedGroupBundle.schoolLevels.every(
-          (l) => l.schoolLevel.id != _selectedSchoolLevelId,
-        )) {
-      _selectedSchoolLevelId = resolvedGroupBundle.schoolLevels.isNotEmpty
-          ? resolvedGroupBundle.schoolLevels.first.schoolLevel.id
-          : '';
-      _initialSchoolLevelId = _selectedSchoolLevelId;
-      bootstrapChanged = true;
-    }
-
-    _bootstrapDefaultsApplied = true;
-
-    if (bootstrapChanged) {
-      WidgetsBinding.instance.addPostFrameCallback((_) {
-        if (!mounted) return;
-        _recomputeFormState();
-      });
-    }
   }
 
   @override
@@ -408,69 +330,58 @@ class PreviousAcademicInfoStepState extends State<PreviousAcademicInfoStep> {
 
     return BlocProvider<EnrollmentAcademicInfoBloc>.value(
       value: _academicInfoBloc,
-      child: BlocConsumer<EnrollmentAcademicInfoBloc, EnrollmentAcademicInfoState>(
-        listenWhen: (prev, curr) => prev.status != curr.status,
-        listener: (context, state) {
-          _onSavingChanged(state.status == EnrollmentAcademicInfoStatus.loading);
+      child:
+          BlocConsumer<EnrollmentAcademicInfoBloc, EnrollmentAcademicInfoState>(
+            listenWhen: (prev, curr) => prev.status != curr.status,
+            listener: (context, state) {
+              _onSavingChanged(
+                state.status == EnrollmentAcademicInfoStatus.loading,
+              );
 
-          if (state.status == EnrollmentAcademicInfoStatus.success) {
-            _markCurrentAsSavedSnapshot();
-            _recomputeFormState();
-            _onSavingChanged(false);
-            if (_showValidationHints) {
-              setState(() => _showValidationHints = false);
-            }
-            context.read<EnrollmentBloc>().add(
-              EnrollmentDetailRequested(
-                enrollmentId: widget.enrollmentId,
-                silent: true,
-              ),
-            );
-            AppSnackBar.showSuccess(context, l10n.academicInfoSaveSuccess);
-          } else if (state.status == EnrollmentAcademicInfoStatus.failure) {
-            _onSavingChanged(false);
-            AppSnackBar.showError(
-              context,
-              l10n.academicInfoSaveError(state.errorMessage ?? ''),
-            );
-          }
-        },
-        builder: (context, state) {
-          final isLoading = state.status == EnrollmentAcademicInfoStatus.loading;
-          return BlocBuilder<BootstrapBloc, BootstrapState>(
-            builder: (context, bootstrapState) {
-              final bootstrap = bootstrapState.bootstrap;
-              if (bootstrap != null) {
-                _applyBootstrapDefaults(bootstrap);
+              if (state.status == EnrollmentAcademicInfoStatus.success) {
+                _markCurrentAsSavedSnapshot();
+                _recomputeFormState();
+                _onSavingChanged(false);
+                if (_showValidationHints) {
+                  setState(() => _showValidationHints = false);
+                }
+                context.read<EnrollmentBloc>().add(
+                  EnrollmentDetailRequested(
+                    enrollmentId: widget.enrollmentId,
+                    silent: true,
+                  ),
+                );
+                AppSnackBar.showSuccess(context, l10n.academicInfoSaveSuccess);
+              } else if (state.status == EnrollmentAcademicInfoStatus.failure) {
+                _onSavingChanged(false);
+                AppSnackBar.showError(
+                  context,
+                  l10n.academicInfoSaveError(state.errorMessage ?? ''),
+                );
               }
-
-              return AcademicInfoStepBody(
-                bootstrap: bootstrap,
+            },
+            builder: (context, state) {
+              final isLoading =
+                  state.status == EnrollmentAcademicInfoStatus.loading;
+              return PreviousAcademicInfoStepBody(
                 prevYearController: _prevYearController,
                 prevSchoolController: _prevSchoolController,
                 prevCycleController: _prevCycleController,
                 prevLevelController: _prevLevelController,
                 prevRateController: _prevRateController,
                 prevRankController: _prevRankController,
-                currYearController: _currYearController,
-                targetOptionController: _targetOptionController,
                 validatedPreviousYear: _validatedPreviousYear,
-                selectedSchoolLevelGroupId: _selectedSchoolLevelGroupId,
-                selectedSchoolLevelId: _selectedSchoolLevelId,
                 showValidation: showValidation,
                 isLoading: isLoading,
                 canSave: _canSave,
                 showInlineSaveButton: widget.showInlineSaveButton,
-                showPreviousSection: true,
-                showTargetSection: false,
                 onSave: _onSave,
                 onValidatedChanged: (value) {
                   setState(() => _validatedPreviousYear = value);
                   _recomputeFormState();
                 },
-                onGroupChanged: (_, __) {},
-                onLevelChanged: (_) {},
-                prevYearError: showValidation && _prevYearController.text.trim().isEmpty
+                prevYearError:
+                    showValidation && _prevYearController.text.trim().isEmpty
                     ? l10n.requiredFieldError(l10n.academicYearLabel)
                     : null,
                 prevSchoolError:
@@ -485,22 +396,24 @@ class PreviousAcademicInfoStepState extends State<PreviousAcademicInfoStep> {
                     showValidation && _prevLevelController.text.trim().isEmpty
                     ? l10n.requiredFieldError(l10n.schoolLevelLabel)
                     : null,
-                prevRateError: showValidation &&
+                prevRateError:
+                    showValidation &&
                         (_prevRateController.text.trim().isEmpty ||
-                            double.tryParse(_prevRateController.text.trim()) == null)
+                            double.tryParse(_prevRateController.text.trim()) ==
+                                null)
                     ? (_prevRateController.text.trim().isEmpty
-                        ? l10n.requiredFieldError(l10n.averageLabel)
-                        : l10n.invalidNumberFieldError(l10n.averageLabel))
+                          ? l10n.requiredFieldError(l10n.averageLabel)
+                          : l10n.invalidNumberFieldError(l10n.averageLabel))
                     : null,
-                prevRankError: showValidation &&
+                prevRankError:
+                    showValidation &&
                         (_prevRankController.text.trim().isEmpty ||
-                            int.tryParse(_prevRankController.text.trim()) == null)
+                            int.tryParse(_prevRankController.text.trim()) ==
+                                null)
                     ? (_prevRankController.text.trim().isEmpty
-                        ? l10n.requiredFieldError(l10n.rankingLabel)
-                        : l10n.invalidNumberFieldError(l10n.rankingLabel))
+                          ? l10n.requiredFieldError(l10n.rankingLabel)
+                          : l10n.invalidNumberFieldError(l10n.rankingLabel))
                     : null,
-                groupError: null,
-                levelError: null,
                 prevYearChanged: _isChanged(
                   _prevYearController.text,
                   _initialPrevYear,
@@ -518,18 +431,16 @@ class PreviousAcademicInfoStepState extends State<PreviousAcademicInfoStep> {
                   _initialPrevLevel,
                 ),
                 prevRateChanged:
-                    _normalizeRate(_prevRateController.text) != _initialPrevRate,
+                    _normalizeRate(_prevRateController.text) !=
+                    _initialPrevRate,
                 prevRankChanged:
-                    _normalizeRank(_prevRankController.text) != _initialPrevRank,
+                    _normalizeRank(_prevRankController.text) !=
+                    _initialPrevRank,
                 validatedPreviousYearChanged:
                     _validatedPreviousYear != _initialValidatedPreviousYear,
-                groupChanged: false,
-                levelChanged: false,
               );
             },
-          );
-        },
-      ),
+          ),
     );
   }
 }
